requirements.txt
fastapi
uvicorn
elasticsearch
loguru
python-dotenvuvicorn==0.34.2
pydantic==2.11.4
pydantic_core==2.33.2
